return back deleted code and extract transitions into functions and outer file
