+++ app/src/androidTest/java/io/github/omisie11/spacexfollower/NavDrawerEspressoTests.kt
@@ -17,6 +17,7 @@ import io.github.omisie11.spacexfollower.ui.MainActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+
 @RunWith(AndroidJUnit4::class)
 @LargeTest
 class NavDrawerEspressoTests {
