+++ app/src/main/java/io/github/omisie11/spacexfollower/ui/cores/CoresFragment.kt
@@ -46,19 +46,7 @@ class CoresFragment : Fragment(), CoresAdapter.OnItemClickListener {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        enterTransition = MaterialContainerTransform().apply {
-            startView = requireActivity().findViewById(R.id.text_label_cores)
-            endView = root
-            duration = resources.getInteger(R.integer.core_motion_duration_large).toLong()
-            scrimColor = Color.TRANSPARENT
-            containerColor = requireContext().themeColor(R.attr.colorSurface)
-            startContainerColor = requireContext().themeColor(R.attr.colorSurface)
-            endContainerColor = requireContext().themeColor(R.attr.colorSurface)
-        }
-        returnTransition = Slide().apply {
-            duration = resources.getInteger(R.integer.core_motion_duration_medium).toLong()
-            addTarget(R.id.root)
-        }
+        enterAndReturnTransition()
         // Setup recyclerView
         viewAdapter = CoresAdapter(this)
         recyclerView.apply {
@@ -146,6 +134,22 @@ class CoresFragment : Fragment(), CoresAdapter.OnItemClickListener {
                 }
             }
         })
+    }
+
+    private fun enterAndReturnTransition() {
+        enterTransition = MaterialContainerTransform().apply {
+            startView = requireActivity().findViewById(R.id.text_label_cores)
+            endView = root
+            duration = resources.getInteger(R.integer.core_motion_duration_large).toLong()
+            scrimColor = Color.TRANSPARENT
+            containerColor = requireContext().themeColor(R.attr.colorSurface)
+            startContainerColor = requireContext().themeColor(R.attr.colorSurface)
+            endContainerColor = requireContext().themeColor(R.attr.colorSurface)
+        }
+        returnTransition = Slide().apply {
+            duration = resources.getInteger(R.integer.core_motion_duration_medium).toLong()
+            addTarget(R.id.root)
+        }
     }
 
     override fun onResume() {
