+++ app/src/main/java/io/github/omisie11/spacexfollower/ui/dashboard/DashboardFragment.kt
@@ -33,7 +33,7 @@ class DashboardFragment : Fragment(R.layout.fragment_dashboard) {
 
     private val viewModel: DashboardViewModel by sharedViewModel()
 
-    val currentNavigationFragment: Fragment?
+    private val currentNavigationFragment: Fragment?
         get() = activity?.supportFragmentManager?.findFragmentById(R.id.nav_host_fragment)
             ?.childFragmentManager
             ?.fragments
@@ -93,42 +93,32 @@ class DashboardFragment : Fragment(R.layout.fragment_dashboard) {
         }
 
         text_label_capsules.setOnClickListener {
-            currentNavigationFragment?.apply {
-                exitTransition = MaterialElevationScale(false).apply {
-                    duration = resources.getInteger(R.integer.core_motion_duration_large).toLong()
-                }
-                reenterTransition = MaterialElevationScale(true).apply {
-                    duration = resources.getInteger(R.integer.core_motion_duration_large).toLong()
-                }
-            }
+            exitAndReEnterTransition()
             findNavController()
                 .navigate(DashboardFragmentDirections.actionDashboardDestToCapsulesDest())
         }
 
         text_label_launches.setOnClickListener {
-            currentNavigationFragment?.apply {
-                exitTransition = MaterialElevationScale(false).apply {
-                    duration = resources.getInteger(R.integer.core_motion_duration_large).toLong()
-                }
-                reenterTransition = MaterialElevationScale(true).apply {
-                    duration = resources.getInteger(R.integer.core_motion_duration_large).toLong()
-                }
-            }
+            exitAndReEnterTransition()
             findNavController()
                 .navigate(DashboardFragmentDirections.actionDashboardDestToLaunchesDest())
         }
 
         text_label_cores.setOnClickListener {
-            currentNavigationFragment?.apply {
-                exitTransition = MaterialElevationScale(false).apply {
-                    duration = resources.getInteger(R.integer.core_motion_duration_large).toLong()
-                }
-                reenterTransition = MaterialElevationScale(true).apply {
-                    duration = resources.getInteger(R.integer.core_motion_duration_large).toLong()
-                }
-            }
+            exitAndReEnterTransition()
             findNavController()
                 .navigate(DashboardFragmentDirections.actionDashboardDestToCoresDest())
+        }
+    }
+
+    private fun exitAndReEnterTransition() {
+        currentNavigationFragment?.apply {
+            exitTransition = MaterialElevationScale(false).apply {
+                duration = resources.getInteger(R.integer.core_motion_duration_large).toLong()
+            }
+            reenterTransition = MaterialElevationScale(true).apply {
+                duration = resources.getInteger(R.integer.core_motion_duration_large).toLong()
+            }
         }
     }
