+++ app/src/main/java/io/github/omisie11/spacexfollower/util/ResourcesExtensions.kt
@@ -1,0 +1,18 @@
+package io.github.omisie11.spacexfollower.util
+
+import android.content.Context
+import android.graphics.Color
+import androidx.annotation.AttrRes
+import androidx.annotation.ColorInt
+import androidx.core.content.res.use
+
+@ColorInt
+fun Context.themeColor(
+    @AttrRes themeAttrId: Int
+): Int {
+    return obtainStyledAttributes(
+        intArrayOf(themeAttrId)
+    ).use {
+        it.getColor(0, Color.MAGENTA)
+    }
+}
