+++ app/src/main/res/layout/activity_main.xml
@@ -34,7 +34,7 @@
                 android:layout_height="match_parent"
                 app:layout_behavior="com.google.android.material.appbar.AppBarLayout$ScrollingViewBehavior">
 
-            <fragment
+            <androidx.fragment.app.FragmentContainerView
                     android:id="@+id/nav_host_fragment"
                     android:name="androidx.navigation.fragment.NavHostFragment"
                     android:layout_width="match_parent"
